+++ src/db/startup.py
@@ -38,7 +38,8 @@ queries = {
         "new_version": "v1.0.0",
     },
     "v1.0.0": {"query": None, "new_version": "v1.1.0"},
-    "v1.1.0": None,
+    "v1.1.0": {"query": None, "new_version": "v1.1.1"},
+    "v1.1.1": None,
 }
 
 create_db = """